Behaviour change: eception instead of crash
The default behaviour of the library is to crash if an assert fails.
This can be overridden by defining your own assert, before including. This
is a bit complicated because has to be done in each file where rapidjsonxml is
included.

This commit changes the default behaviour to throw an exception instead of
crashing.

It uses define, so if client code defined their own assert, they will use
that one instead of the one generating the exception. The behaviour changed is
the default one, not any user defined one.